requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.6.2
    # via httpcore
beautifulsoup4==4.11.2
    # via -r requirements.in
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
h11==0.14.0
    # via httpcore
httpcore==0.16.3
    # via httpx
httpx==0.23.3
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   rfc3986
loguru==0.6.0
    # via -r requirements.in
mypy==1.0.1
    # via -r requirements.in
mypy-extensions==1.0.0
    # via mypy
pydub==0.25.1
    # via -r requirements.in
pydub-stubs==0.25.1.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
rfc3986[idna2008]==1.5.0
    # via httpx
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
soupsieve==2.4
    # via beautifulsoup4
tomli==2.0.1
    # via mypy
types-beautifulsoup4==4.11.6.7
    # via -r requirements.in
types-html5lib==1.1.11.12
    # via types-beautifulsoup4
typing-extensions==4.5.0
    # via mypy
